backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
et-xmlfile==2.0.0
openpyxl==3.1.5
Pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
